+++ Core/DbParameters.cs
@@ -173,7 +173,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public DbParameters Put<T>(string name, T v, ushort x = 0) where T : IPersist
+        public DbParameters Put<T>(string name, T v, uint x = 0) where T : IPersist
         {
             if (name == null)
             {
@@ -288,7 +288,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public DbParameters Put<T>(string name, T[] v, ushort x = 0) where T : IPersist
+        public DbParameters Put<T>(string name, T[] v, uint x = 0) where T : IPersist
         {
             if (name == null)
             {
