+++ Core/WebContext.cs
@@ -108,7 +108,7 @@ namespace Greatbone.Core
             }
         }
 
-        public T Obj<T>(ushort x = 0) where T : IPersist, new()
+        public T Obj<T>(uint x = 0) where T : IPersist, new()
         {
             JObj jo = JObj;
             T obj = new T();
@@ -116,7 +116,7 @@ namespace Greatbone.Core
             return obj;
         }
 
-        public T[] Arr<T>(ushort x = 0) where T : IPersist, new()
+        public T[] Arr<T>(uint x = 0) where T : IPersist, new()
         {
             JArr ja = JArr;
             T[] arr = new T[ja.Count];
@@ -269,7 +269,7 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public bool Got<V>(string name, ref V v, ushort x = 0) where V : IPersist, new()
+        public bool Got<V>(string name, ref V v, uint x = 0) where V : IPersist, new()
         {
             StringValues values;
             if (Request.Query.TryGetValue(name, out values))
@@ -389,7 +389,7 @@ namespace Greatbone.Core
             return false;
         }
 
-        public bool Got<V>(string name, ref V[] v, ushort x = 0) where V : IPersist, new()
+        public bool Got<V>(string name, ref V[] v, uint x = 0) where V : IPersist, new()
         {
             StringValues values;
             if (Request.Query.TryGetValue(name, out values))
@@ -433,12 +433,12 @@ namespace Greatbone.Core
 
         internal int MaxAge { get; set; }
 
-        public void Respond<T>(int status, T obj, ushort x = 0, bool? pub = false, int maxage = 0) where T : IPersist
+        public void Respond<T>(int status, T obj, uint x = 0, bool? pub = false, int maxage = 0) where T : IPersist
         {
             Respond(status, jcont => jcont.PutObj(obj, x), pub, maxage);
         }
 
-        public void Respond<T>(int status, T[] arr, ushort x = 0, bool? pub = false, int maxage = 0) where T : IPersist
+        public void Respond<T>(int status, T[] arr, uint x = 0, bool? pub = false, int maxage = 0) where T : IPersist
         {
             Respond(status, jcont => jcont.PutArr(arr, x), pub, maxage);
         }
