+++ Sample/Fame.cs
@@ -31,7 +31,7 @@ namespace Greatbone.Sample
         internal Ref[] friends;
         internal Ref[] awards;
 
-        public void Load(ISource sc, ushort x = 0)
+        public void Load(ISource sc, uint x = 0)
         {
             sc.Got(nameof(id), ref id);
             sc.Got(nameof(name), ref name);
@@ -59,7 +59,7 @@ namespace Greatbone.Sample
             sc.Got(nameof(awards), ref awards);
         }
 
-        public void Save<R>(ISink<R> sk, ushort x = 0) where R : ISink<R>
+        public void Save<R>(ISink<R> sk, uint x = 0) where R : ISink<R>
         {
             sk.Put(nameof(id), id);
             sk.Put(nameof(name), name);
@@ -97,14 +97,14 @@ namespace Greatbone.Sample
 
         internal string hint;
 
-        public void Load(ISource sc, ushort x = 0)
+        public void Load(ISource sc, uint x = 0)
         {
             sc.Got(nameof(name), ref name);
             sc.Got(nameof(@ref), ref @ref);
             sc.Got(nameof(hint), ref hint);
         }
 
-        public void Save<R>(ISink<R> sk, ushort x = 0) where R : ISink<R>
+        public void Save<R>(ISink<R> sk, uint x = 0) where R : ISink<R>
         {
             sk.Put(nameof(name), name);
             sk.Put(nameof(@ref), @ref);
